requirements.txt
Flask
email_validator
Flask-MySQL
Flask-WTF
jinja2~=3.0.3
Werkzeug~=2.0.0
markupsafe==2.0.1
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-exporter-otlp